Fix Metalinks... I Think
A preallocation issue prevented metalink files from being read. Basically, the size of the intended file was preallocated and then filled with the metalink XML data, leaving a bunch of null bytes at the end of the file which then triggered an error when trying to get responseXML from an XHR read of the file. The real size is eventually passed along, but as the state moves on from RUNNING to FINISHING by that point, the re-preallocation never occurs. So I added a check for FINISHING isMetalink file, and then call the preallocator immediately, just to be sure.